Create Maze_create&show
I have found the solution I could understand on how we could create a maze and show it. The problem is that it looks like it is not read inline and it is automatically shown. But we need the one that is shown when "show is typed". Hope it will still be hepfull for something.